# internal/delivery/http/router.go
package http

import (
	"context"
	"log/slog"
	"net/http"
	"time"

	"Selecto-Ecommerce/internal/config"
	"Selecto-Ecommerce/internal/delivery/http/handlers"
	"Selecto-Ecommerce/internal/delivery/http/middleware"
	"Selecto-Ecommerce/internal/infrastructure/database"
	mailinfra "Selecto-Ecommerce/internal/infrastructure/email"
	"Selecto-Ecommerce/internal/jobs"
	"Selecto-Ecommerce/internal/repository/postgres"
	"Selecto-Ecommerce/internal/shared/logging"

	"github.com/gin-contrib/cors"
	"github.com/gin-gonic/gin"
)

func SetupRouter(
	db *database.DB,
	cfg *config.Config,
	logger *slog.Logger,
	notifiers ...mailinfra.DispatchNotifier,
) *gin.Engine {
	if cfg.AppEnv == "production" {
		gin.SetMode(gin.ReleaseMode)
	}
	r := gin.New()
	r.Use(gin.Recovery())
	_ = r.SetTrustedProxies([]string{"127.0.0.1", "::1"})
	r.Use(middleware.RequestContext(logger))
	r.Use(middleware.SecurityHeaders(cfg.AppEnv))
	r.Use(middleware.RequestBodyLimit(1 << 20))
	r.Use(middleware.RateLimit(cfg.RateLimitPerMinute))
	r.Use(cors.New(corsConfig(cfg)))

	r.GET("/health", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{"status": "ok"})
	})
	r.GET("/ready", func(c *gin.Context) {
		ctx, cancel := context.WithTimeout(c.Request.Context(), 2*time.Second)
		defer cancel()
		if err := db.Pool.Ping(ctx); err != nil {
			c.JSON(http.StatusServiceUnavailable, gin.H{"status": "not_ready"})
			return
		}
		c.JSON(http.StatusOK, gin.H{"status": "ready"})
	})

	r.POST("/register", handlers.RegisterHandler(db, cfg, logger, notifiers...))
	r.GET("/seo/head", handlers.SEOHeadHandler(db, cfg))
	r.GET("/sitemap.xml", handlers.SitemapHandler(db, cfg))
	r.POST("/login", handlers.LoginHandler(db, cfg, logger))
	r.POST("/auth/email/verify", handlers.VerifyEmailHandler(db))
	r.POST("/auth/email/resend", handlers.ResendVerificationHandler(db, cfg, notifiers...))
	r.POST("/auth/password/forgot", handlers.ForgotPasswordHandler(db, cfg, notifiers...))
	r.POST("/auth/password/reset", handlers.ResetPasswordHandler(db))
	r.POST("/admin/auth/password/forgot", handlers.AdminForgotPasswordHandler(db, cfg, notifiers...))
	r.POST("/admin/auth/password/reset", handlers.AdminResetPasswordHandler(db))
	r.POST("/auth/google", handlers.GoogleAuthHandler(db, cfg, logger, nil))
	r.POST("/auth/google/register", handlers.GoogleRegisterHandler(db, cfg, logger))
	r.GET("/products", handlers.GetProductsHandler(db, logger))
	r.GET("/categories", handlers.ListCategoriesHandler(db, true))
	r.GET("/product-images/:id", handlers.ProductImageHandler(db))
	r.GET("/products/:id/reviews", handlers.ListProductReviewsHandler(db))
	carousel := postgres.CarouselStore{Pool: db.Pool}
	r.GET("/carousel-slides", handlers.ListCarouselHandler(carousel, true))
	r.GET("/carousel-images/:id", handlers.CarouselImageHandler(carousel, true))
	marketingNotifiers := []mailinfra.MarketingDispatchNotifier{}
	if cfg.BrevoContactsEnabled {
		for _, notifier := range notifiers {
			if marketingNotifier, ok := notifier.(mailinfra.MarketingDispatchNotifier); ok {
				marketingNotifiers = append(marketingNotifiers, marketingNotifier)
			}
		}
	}
	r.POST("/marketing/newsletter", handlers.NewsletterSubscribeHandler(db, marketingNotifiers...))
	r.POST("/marketing/newsletter/unsubscribe", handlers.NewsletterUnsubscribeHandler(db, cfg, notifiers...))
	r.POST("/marketing/newsletter/unsubscribe/confirm", handlers.NewsletterUnsubscribeConfirmHandler(db, marketingNotifiers...))
	r.POST("/communications/brevo/webhook", handlers.BrevoWebhookHandler(db, cfg))
	r.POST("/payments/webhook", middleware.InternalWebhookAuth(cfg.InternalWebhookSecret, 5*time.Minute), handlers.PaymentWebhookHandler(db, cfg, logger, notifiers...))

	authorized := r.Group("/")
	authorized.Use(middleware.AuthMiddleware(db, cfg.JWTSecret))

	authorized.GET("/me", handlers.GetMeHandler(db))
	authorized.PATCH("/me", handlers.UpdateMeHandler(db))

	authorized.POST("/orders", handlers.CreateOrderHandler(db, cfg, logger, notifiers...))
	authorized.GET("/orders/:id", handlers.GetOrderHandler(db))
	authorized.POST("/orders/:id/cancel", handlers.CancelOrderHandler(db, cfg, logger, notifiers...))
	authorized.GET("/my-orders", handlers.GetMyOrdersHandler(db))
	authorized.POST("/checkout", handlers.CheckoutHandler(db, cfg, logger))
	authorized.POST("/products/:id/reviews", handlers.SubmitProductReviewHandler(db))
	authorized.POST("/auth/google/link", handlers.GoogleLinkHandler(db, cfg, logger, nil))

	admin := authorized.Group("/")
	admin.Use(middleware.RequireAdmin(db, logger))

	admin.GET("/admin/me", handlers.GetAdminMeHandler(db))
	admin.GET("/admin/dashboard", handlers.GetAdminDashboardHandler(db, logger))
	admin.GET("/admin/categories", handlers.ListCategoriesHandler(db, false))
	admin.POST("/admin/categories", handlers.CreateCategoryHandler(db))
	admin.PATCH("/admin/categories/reorder", handlers.ReorderCategoriesHandler(db))
	admin.PATCH("/admin/categories/:id", handlers.UpdateCategoryHandler(db))
	admin.GET("/admin/carousel-slides", handlers.ListCarouselHandler(carousel, false))
	admin.POST("/admin/carousel-slides", handlers.SaveCarouselHandler(carousel, true))
	admin.PATCH("/admin/carousel-slides/:id", handlers.SaveCarouselHandler(carousel, false))
	admin.DELETE("/admin/carousel-slides/:id", handlers.DeleteCarouselHandler(carousel))
	admin.GET("/admin/carousel-slides/:id/image", handlers.CarouselImageHandler(carousel, false))
	admin.GET("/admin/products", handlers.AdminListProductsHandler(db))
	admin.GET("/admin/products/:id", handlers.AdminGetProductHandler(db))
	admin.POST("/admin/products", handlers.AdminCreateProductHandler(db, logger))
	admin.PATCH("/admin/products/:id", handlers.AdminUpdateProductHandler(db, logger))
	admin.PATCH("/admin/products/:id/status", handlers.AdminUpdateProductStatusHandler(db, logger))
	admin.DELETE("/admin/products/:id", handlers.AdminDeleteProductHandler(db, logger))
	admin.POST("/admin/products/:id/stock-adjustments", handlers.AdminAdjustProductStockHandler(db, logger))
	admin.POST("/admin/products/:id/images", handlers.AdminUploadProductImageHandler(db))
	admin.PATCH("/admin/products/:id/images/:image_id", handlers.AdminUpdateProductImageHandler(db))
	admin.DELETE("/admin/products/:id/images/:image_id", handlers.AdminDeleteProductImageHandler(db))
	admin.GET("/admin/product-reviews", handlers.AdminOperationsListProductReviewsHandler(db))
	admin.PATCH("/admin/product-reviews/:id", handlers.AdminOperationsModerateProductReviewHandler(db))
	admin.GET("/admin/orders", handlers.AdminListOrdersHandler(db))
	admin.GET("/admin/orders/:id", handlers.AdminGetOrderHandler(db))
	admin.POST("/admin/orders/:id/cancel", handlers.AdminCancelOrderHandler(db, logger))
	admin.PATCH("/admin/orders/:id/shipment", handlers.AdminUpdateShipmentHandler(db, cfg, logger, notifiers...))
	admin.GET("/admin/customers", handlers.AdminListCustomersHandler(db))
	admin.GET("/admin/customers/:id", handlers.AdminGetCustomerHandler(db))
	admin.GET("/admin/marketing/subscriptions", handlers.AdminOperationsListMarketingSubscriptionsHandler(db))
	admin.GET("/admin/communications/outbox", handlers.AdminOperationsListEmailOutboxHandler(db))
	admin.GET("/admin/communications/brevo/events", handlers.AdminListBrevoEventsHandler(db))
	admin.GET("/admin/communications/brevo/status", handlers.AdminBrevoStatusHandler(db, cfg))
	admin.POST("/admin/marketing/subscriptions/:id/sync", handlers.AdminRetryMarketingSyncHandler(db, cfg, marketingNotifiers...))
	admin.GET("/admin/audit-logs", handlers.AdminListAuditLogsHandler(db))
	admin.GET("/admin/audit-logs/entities/:entity_type/:entity_id", handlers.AdminListEntityAuditLogsHandler(db))
	admin.GET("/admin/observability", handlers.AdminObservabilityHandler(db, cfg))
	admin.GET("/admin/payments", handlers.AdminPaymentsProxyHandler(cfg, "/internal/admin/payments"))
	admin.GET("/admin/payments/records/:record_id", handlers.AdminPaymentsProxyHandler(cfg, "/internal/admin/payments/records/:record_id"))
	admin.GET("/admin/payments/:payment_id", handlers.AdminPaymentsProxyHandler(cfg, "/internal/admin/payments/:payment_id"))
	admin.GET("/admin/payment-preferences", handlers.AdminPaymentsProxyHandler(cfg, "/internal/admin/payment-preferences"))
	admin.GET("/admin/payment-webhooks", handlers.AdminPaymentsProxyHandler(cfg, "/internal/admin/webhook-events"))
	admin.GET("/admin/payment-notifications", handlers.AdminPaymentsProxyHandler(cfg, "/internal/admin/ecommerce-notifications"))
	admin.GET("/admin/payments-metrics", handlers.AdminPaymentsProxyHandler(cfg, "/internal/admin/metrics"))

	admin.POST("/products", handlers.CreateProductHandler(db, logger))
	admin.GET("/admin/metrics", handlers.GetMetricsHandler(db, logger))
	admin.GET("/admin/test", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{
			"message": "welcome admin",
		})
	})

	return r
}

func corsConfig(cfg *config.Config) cors.Config {
	corsCfg := cors.Config{
		AllowOrigins:  cfg.CORSAllowedOrigins,
		AllowMethods:  []string{"GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"},
		AllowHeaders:  []string{"Origin", "Content-Type", "Accept", "Authorization", "X-Request-ID", "X-Correlation-ID", "X-Service-Name", "X-Service-Timestamp", "X-Service-Signature", "Idempotency-Key", "X-Selecto-Signature", "X-Selecto-Timestamp", "X-Destry-Signature", "X-Destry-Timestamp"},
		ExposeHeaders: []string{"X-Request-ID", "X-Correlation-ID"},
	}

	if cfg.AppEnv != "production" {
		corsCfg.AllowOrigins = nil
		corsCfg.AllowAllOrigins = true
	}

	return corsCfg
}

func StartExpiredOrderWorker(ctx context.Context, db *database.DB, cfg *config.Config, logger *slog.Logger) {
	go func() {
		process := func() {
			if err := jobs.Run(ctx, jobs.ExpireOrders, db, cfg, logger); err != nil {
				logger.Error(logging.EventExpiredOrderReleaseFailed, "error", err)
			}
		}
		process()
		ticker := time.NewTicker(cfg.ReleaseWorkerInterval)
		defer ticker.Stop()
		for {
			select {
			case <-ctx.Done():
				return
			case <-ticker.C:
				process()
			}
		}
	}()
}