# ops/api-gateway/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Selecto Ecommerce API sobre infraestructura Destry",
    "version": "corte-selecto-v1"
  },
  "schemes": ["https"],
  "produces": ["application/json"],
  "consumes": ["application/json"],
...
  "paths": {
    "/seo/head": {
      "get": {"operationId":"PublicSEOHead","produces":["text/html"],"parameters":[{"name":"path","in":"query","type":"string"}],"responses":{"200":{"description":"Escaped public metadata fragment"}}},
      "options":{"operationId":"CorsPublicSEOHead","responses":{"204":{"description":"Preflight"}}}
    },
...
    "/admin/categories/{id}": {
      "patch": {
        "operationId": "AdminUpdateCategory",
        "responses": {
          "200": {
            "description": "AdminUpdateCategory"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "CorsAdminUpdateCategory",
        "responses": {
          "204": {
            "description": "Preflight"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/admin/marketing/subscriptions/{id}/sync": {
      "post": {
        "operationId": "AdminRetryMarketingSync",
        "responses": {
          "202": {
            "description": "AdminRetryMarketingSync"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "CorsAdminRetryMarketingSync",
        "responses": {
          "204": {
            "description": "Preflight"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/carousel-images/{id}": {
      "parameters":[{"name":"id","in":"path","required":true,"type":"integer"}],
      "get": {"operationId":"CarouselImage","responses":{"200":{"description":"Imagen publicada"},"404":{"description":"No publicada o destino no disponible"}}},
      "options": {"operationId":"CorsCarouselImage","responses":{"204":{"description":"Preflight"}}}
    },
    "/admin/carousel-slides": {
      "get": {"operationId":"AdminListCarousel","responses":{"200":{"description":"Diapositivas, máximo 20"}}},
      "post": {"operationId":"AdminCreateCarousel","consumes":["multipart/form-data"],"responses":{"201":{"description":"Diapositiva creada"},"409":{"description":"Límite o destino no disponible"}}},
      "options": {"operationId":"CorsAdminCarousel","responses":{"204":{"description":"Preflight"}}}
    },
    "/admin/carousel-slides/{id}": {
      "parameters":[{"name":"id","in":"path","required":true,"type":"integer"}],
      "patch": {"operationId":"AdminUpdateCarousel","consumes":["multipart/form-data"],"responses":{"200":{"description":"Diapositiva actualizada"},"409":{"description":"Versión o destino no disponible"}}},
      "delete": {"operationId":"AdminDeleteCarousel","responses":{"204":{"description":"Diapositiva eliminada"},"409":{"description":"Versión cambió"}}},
      "options": {"operationId":"CorsAdminCarouselItem","responses":{"204":{"description":"Preflight"}}}
    },
    "/admin/carousel-slides/{id}/image": {
      "parameters":[{"name":"id","in":"path","required":true,"type":"integer"}],
      "get": {"operationId":"AdminCarouselImage","responses":{"200":{"description":"Vista previa autenticada"}}},
      "options": {"operationId":"CorsAdminCarouselPreview","responses":{"204":{"description":"Preflight"}}}
    },
...
    "/admin/categories": {
      "get": {"operationId":"AdminListCategories","responses":{"200":{"description":"Categorías paginadas"}}},
      "post": {"operationId":"AdminCreateCategory","responses":{"201":{"description":"Categoría creada"},"409":{"description":"Nombre o slug duplicado"}}},
      "options": {"operationId":"CorsAdminCategories","responses":{"204":{"description":"Preflight"}}}
    },
...
    "/admin/audit-logs/entities/{entity_type}/{entity_id}": {
      "get": {
        "operationId": "AdminListEntityAuditLogs",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_35",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/admin/customers/{id}": {
      "get": {
        "operationId": "AdminGetCustomer",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_33",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/admin/orders/{id}": {
      "get": {
        "operationId": "AdminGetOrder",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_29",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/orders/{id}/cancel": {
      "post": {
        "operationId": "AdminCancelOrder",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_30",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/orders/{id}/shipment": {
      "patch": {
        "operationId": "AdminUpdateShipment",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_31",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/admin/payments/records/{record_id}": {
      "get": {
        "operationId": "AdminGetPaymentRecord",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_admin_payment_record",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/payments/{payment_id}": {
      "get": {
        "operationId": "AdminGetPayment",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_38",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/admin/product-reviews/{id}": {
      "patch": {
        "operationId": "AdminModerateProductReview",
        "responses": { "200": { "description": "Reseña moderada" } },
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }]
      },
      "options": {
        "operationId": "CorsAdminProductReview",
        "responses": { "204": { "description": "Preflight" } },
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }]
      }
    },
...
    "/admin/products/{id}": {
      "get": {
        "operationId": "AdminGetProduct",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "patch": {
        "operationId": "AdminUpdateProduct",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "delete": {
        "operationId": "AdminDeleteProduct",
        "responses": {
          "204": {
            "description": "Producto eliminado definitivamente"
          },
          "409": { "description": "Conflicto operativo" }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_23",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/products/{id}/images": {
      "post": {
        "operationId": "AdminUploadProductImage",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_26",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/products/{id}/images/{image_id}": {
      "delete": {
        "operationId": "AdminDeleteProductImage",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "patch": {
        "operationId": "AdminUpdateProductImage",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_27",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/products/{id}/status": {
      "patch": {
        "operationId": "AdminUpdateProductStatus",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_24",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/products/{id}/stock-adjustments": {
      "post": {
        "operationId": "AdminAdjustProductStock",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_25",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/orders/{id}": {
      "get": {
        "operationId": "GetOrder",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_16",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/orders/{id}/cancel": {
      "post": {
        "operationId": "CancelOrder",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_17",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/product-images/{id}": {
      "get": {
        "operationId": "GetProductImage",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      },
      "options": {
        "operationId": "Cors_preflight_13",
        "responses": {
          "200": {
            "description": "Respuesta del servicio"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/products/{id}/reviews": {
      "get": {
        "operationId": "ListProductReviews",
        "responses": { "200": { "description": "Reseñas publicadas" } },
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }]
      },
      "post": {
        "operationId": "SubmitProductReview",
        "responses": { "201": { "description": "Reseña enviada a moderación" } },
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }]
      },
      "options": {
        "operationId": "CorsProductReviews",
        "responses": { "204": { "description": "Preflight" } },
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }]
      }
    },
...
  }
}